Gives the divisors of a user enter number
Using loops, branching and arithmetic instructions in MIPS

Write a MIPS assembly program that prompts the user for a positive integer. The program should then display all the positive factor of the integer and state whether or not the number is prime. You should print an error message if the user does not enter a positive integer. You are not required to have a loop for multiple runs, but I appreciate it when you do. If you use such a loop, you should terminate the loop with the user enters 0 for the number of lights.

Possible sample output:

Please enter a positive integer.
-4
This is not a positive integer. Please try again.

Please enter a positive integer.
24
The factors are:
1
2
3
4
6
8
12
24
The number is not prime

Please enter a positive integer.
23
The factors are:
1
23
The number is prime

Please enter a positive integer.
0
Have a nice day.

Email me your MIPS file.